Refactor Token entity handler
